requirements.txt
discord.py==2.4.0
e621-stable==1.0.4
rule34Py==2.1.0
yt_dlp==2025.1.26
PyNaCl==1.5.0
